requirements.txt
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
xgboost
textblob
python-dotenv
pydantic
